+++ src/document/entities/document.entity.ts
@@ -1,3 +1,4 @@
+import { ApiHideProperty } from '@nestjs/swagger';
 import { Exclude } from 'class-transformer';
 import { DocumentModel } from 'prisma/generated/models';
 
@@ -5,6 +6,7 @@ export class Document implements DocumentModel {
   id: number;
   title: string;
   @Exclude()
+  @ApiHideProperty()
   content: Uint8Array<ArrayBuffer>;
   uploadedAt: Date;
   uploader: string;