+++ src/transaction-forward/transaction-forward.controller.ts
@@ -10,7 +10,8 @@ import {
 import { TransactionForwardService } from './transaction-forward.service';
 import { CreateTransactionForwardDto } from './dto/create-transaction-forward.dto';
 import { UpdateTransactionForwardDto } from './dto/update-transaction-forward.dto';
-import { ApiOperation, ApiTags } from '@nestjs/swagger';
+import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
+import { TransactionForward } from './entities/transaction-forward.entity';
 
 @ApiTags('Transaction Forwards')
 @Controller('transaction/:transactionId/forward')
@@ -21,6 +22,7 @@ export class TransactionForwardController {
 
   @Post()
   @ApiOperation({ summary: 'Forward a transaction' })
+  @ApiResponse({ status: 201, type: TransactionForward })
   create(
     @Param('transactionId') transactionId: string,
     @Body() createTransactionForwardDto: CreateTransactionForwardDto,
@@ -33,12 +35,14 @@ export class TransactionForwardController {
 
   @Get()
   @ApiOperation({ summary: "Get all transaction's forwards" })
+  @ApiResponse({ status: 200, type: [TransactionForward] })
   findAll(@Param('transactionId') transactionId: string) {
     return this.transactionForwardService.findAll(+transactionId);
   }
 
   @Patch(':id')
   @ApiOperation({ summary: "Respond to a transaction's forward" })
+  @ApiResponse({ status: 200, type: TransactionForward })
   update(
     @Param('transactionId') transactionId: string,
     @Param('id') id: string,
@@ -53,6 +57,7 @@ export class TransactionForwardController {
 
   @Delete(':id')
   @ApiOperation({ summary: "Undo a transaction's forward" })
+  @ApiResponse({ status: 200, type: TransactionForward })
   remove(
     @Param('transactionId') transactionId: string,
     @Param('id') id: string,