+++ src/transaction-forward/transaction-forward.service.ts
@@ -1,6 +1,18 @@
 import { Injectable } from '@nestjs/common';
 import { CreateTransactionForwardDto } from './dto/create-transaction-forward.dto';
 import { UpdateTransactionForwardDto } from './dto/update-transaction-forward.dto';
+import { TransactionForward } from './entities/transaction-forward.entity';
+import { TransactionForwardStatus } from 'prisma/generated/enums';
+
+const transactionForward = new TransactionForward();
+transactionForward.receiver = 'Yusuf';
+transactionForward.sender = 'AlMahllawi';
+transactionForward.status = TransactionForwardStatus.WAITING;
+transactionForward.comment = 'Please review';
+transactionForward.seen = false;
+transactionForward.forwardedAt = new Date();
+transactionForward.updatedAt = new Date();
+transactionForward.transactionId = 1;
 
 @Injectable()
 export class TransactionForwardService {
@@ -8,11 +20,17 @@ export class TransactionForwardService {
     transactionId: number,
     createTransactionForwardDto: CreateTransactionForwardDto,
   ) {
-    return `This action adds a new transaction forward to transaction #${transactionId}`;
+    transactionForward.id = 1;
+    transactionForward.status = TransactionForwardStatus.WAITING;
+    transactionForward.receiver = createTransactionForwardDto.receiverName;
+    transactionForward.comment = null;
+    transactionForward.transactionId = transactionId;
+    return transactionForward;
   }
 
   findAll(transactionId: number) {
-    return `This action returns all transaction #${transactionId}'s forwards`;
+    transactionForward.transactionId = transactionId;
+    return [transactionForward];
   }
 
   update(
@@ -20,10 +38,18 @@ export class TransactionForwardService {
     id: number,
     updateTransactionForwardDto: UpdateTransactionForwardDto,
   ) {
-    return `This action updates the transaction #${transactionId}'s forward #${id}`;
+    transactionForward.id = id;
+    transactionForward.status = updateTransactionForwardDto.status;
+    transactionForward.comment = updateTransactionForwardDto.comment;
+    transactionForward.seen = true;
+    transactionForward.updatedAt = new Date();
+    transactionForward.transactionId = transactionId;
+    return transactionForward;
   }
 
   remove(transactionId: number, id: number) {
-    return `This action removes the transaction #${transactionId}'s forward #${id}`;
+    transactionForward.id = id;
+    transactionForward.transactionId = transactionId;
+    return transactionForward;
   }
 }