+++ src/transaction/dto/update-transaction.dto.ts
@@ -1,6 +1,9 @@
 import { PartialType } from '@nestjs/swagger';
 import { CreateTransactionDto } from './create-transaction.dto';
+import { IsBoolean, IsOptional } from 'class-validator';
 
 export class UpdateTransactionDto extends PartialType(CreateTransactionDto) {
+  @IsOptional()
+  @IsBoolean()
   fulfilled?: boolean;
 }