+++ src/transaction/transaction.controller.ts
@@ -6,11 +6,14 @@ import {
   Patch,
   Param,
   Delete,
+  HttpCode,
+  HttpStatus,
 } from '@nestjs/common';
 import { TransactionService } from './transaction.service';
 import { CreateTransactionDto } from './dto/create-transaction.dto';
 import { UpdateTransactionDto } from './dto/update-transaction.dto';
-import { ApiOperation, ApiTags } from '@nestjs/swagger';
+import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
+import { Transaction } from './entities/transaction.entity';
 
 @ApiTags('Transactions')
 @Controller('transactions')
@@ -19,24 +22,28 @@ export class TransactionController {
 
   @Post()
   @ApiOperation({ summary: 'Create a transaction' })
+  @ApiResponse({ status: 201, type: Transaction })
   create(@Body() createTransactionDto: CreateTransactionDto) {
     return this.transactionService.create(createTransactionDto);
   }
 
   @Get()
   @ApiOperation({ summary: 'Get all transactions' })
+  @ApiResponse({ status: 200, type: [Transaction] })
   findAll() {
     return this.transactionService.findAll();
   }
 
   @Get(':id')
   @ApiOperation({ summary: 'Get a transaction by ID' })
+  @ApiResponse({ status: 200, type: Transaction })
   findOne(@Param('id') id: string) {
     return this.transactionService.findOne(+id);
   }
 
   @Patch(':id')
   @ApiOperation({ summary: 'Update a transaction by ID' })
+  @ApiResponse({ status: 200, type: Transaction })
   update(
     @Param('id') id: string,
     @Body() updateTransactionDto: UpdateTransactionDto,
@@ -46,12 +53,15 @@ export class TransactionController {
 
   @Delete(':id')
   @ApiOperation({ summary: 'Delete a transaction by ID' })
+  @ApiResponse({ status: 200, type: Transaction })
   remove(@Param('id') id: string) {
     return this.transactionService.remove(+id);
   }
 
   @Post(':id/document/:documentId')
+  @HttpCode(HttpStatus.OK)
   @ApiOperation({ summary: 'Attach a document to a transaction' })
+  @ApiResponse({ status: 200, type: Transaction })
   attachDocument(
     @Param('id') id: string,
     @Param('documentId') documentId: string,
@@ -61,6 +71,7 @@ export class TransactionController {
 
   @Delete(':id/document/:documentId')
   @ApiOperation({ summary: 'Detach a document from a transaction' })
+  @ApiResponse({ status: 200, type: Transaction })
   detachDocument(
     @Param('id') id: string,
     @Param('documentId') documentId: string,