+++ src/transaction/transaction.service.ts
@@ -1,34 +1,72 @@
 import { Injectable } from '@nestjs/common';
 import { CreateTransactionDto } from './dto/create-transaction.dto';
 import { UpdateTransactionDto } from './dto/update-transaction.dto';
+import { Transaction } from './entities/transaction.entity';
+import { TransactionPriority } from 'prisma/generated/enums';
+
+const transaction = new Transaction();
+transaction.title = 'Declarative Heading';
+transaction.description = 'Descriptive sentence.';
+transaction.type = 'Financial';
+transaction.creator = 'AlMahllawi';
+transaction.createdAt = new Date();
+transaction.fulfilled = false;
+transaction.priority = TransactionPriority.MEDIUM;
 
 @Injectable()
 export class TransactionService {
   create(createTransactionDto: CreateTransactionDto) {
-    return 'This action adds a new transaction';
+    transaction.id = 1;
+    transaction.title = createTransactionDto.title;
+    transaction.description = createTransactionDto.description;
+    transaction.type = createTransactionDto.typeName;
+    transaction.priority =
+      createTransactionDto.priority || TransactionPriority.LOW;
+    transaction.createdAt = new Date();
+    return transaction;
   }
 
   findAll() {
-    return `This action returns all transaction`;
+    // TODO: query parameters
+    return [transaction];
   }
 
   findOne(id: number) {
-    return `This action returns a #${id} transaction`;
+    transaction.id = id;
+    return transaction;
   }
 
   update(id: number, updateTransactionDto: UpdateTransactionDto) {
-    return `This action updates a #${id} transaction`;
+    if (updateTransactionDto.title)
+      transaction.title = updateTransactionDto.title;
+    if (updateTransactionDto.description)
+      transaction.description = updateTransactionDto.description;
+    if (updateTransactionDto.typeName)
+      transaction.type = updateTransactionDto.typeName;
+    if (updateTransactionDto.priority)
+      transaction.priority = updateTransactionDto.priority;
+    if (updateTransactionDto.fulfilled)
+      transaction.fulfilled = updateTransactionDto.fulfilled;
+    return transaction;
   }
 
   remove(id: number) {
-    return `This action removes a #${id} transaction`;
+    transaction.id = id;
+    return transaction;
   }
 
   attachDocument(transactionId: number, documentId: number) {
-    return `This action attaches document #${documentId} to transaction #${transactionId}`;
+    transaction.id = transactionId;
+    transaction.description += `\nAttached Document #${documentId}`;
+    return transaction;
   }
 
   detachDocument(transactionId: number, documentId: number) {
-    return `This action detaches document #${documentId} from transaction #${transactionId}`;
+    transaction.id = transactionId;
+    transaction.description = transaction.description.replace(
+      `\nAttached Document #${documentId}`,
+      '',
+    );
+    return transaction;
   }
 }