+++ src/user/entities/user.entity.ts
@@ -1,9 +1,12 @@
 import { UserModel } from 'prisma/generated/models';
 import { UserGroups } from 'prisma/generated/enums';
-import { ApiProperty } from '@nestjs/swagger';
+import { ApiHideProperty, ApiProperty } from '@nestjs/swagger';
+import { Exclude } from 'class-transformer';
 
 export class User implements UserModel {
   name: string;
+  @Exclude()
+  @ApiHideProperty()
   hashedPassword: string;
   createdAt: Date;
   active: boolean;